feat: added llm tracing
