+++ pyproject.toml
@@ -5,11 +5,14 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
+    "arize-phoenix-otel>=0.9.2",
+    "arize-phoenix[evals]>=9.5.0",
     "docker>=6.0.0",
     "gradio>=4.0",
     "langchain-openai>=0.3.16",
     "langchain>=0.0.200",
     "langgraph>=0.0.23",
+    "openinference-instrumentation-langchain>=0.1.42",
     "python-dotenv>=1.1.0",
     "python-nmap>=0.7.1",
 ]
